Make it explicitly permissible to register the empty unit.
Previously we rejected that in registerUnit, but it could still be registered by
stealth by calling getUnitId with the empty string, in which the []
dereference would create a record with a duplicated id of 0. We've been
accidentally relying on that all along, to supply the empty unit in unit
selection dropdowns, so let's properly support it and remove the
unpredictable and apparently-wrong behaviour.